fix for spring cloud test
