+++ .gitignore
@@ -24,6 +24,7 @@ coverage.txt
 
 *.exe
 test
-
+!samples/springcloud/zookeeper/test
+!samples/springcloud/nacos/test
 
 target/
