+++ samples/springcloud/nacos/pixiu/conf.yaml
