+++ samples/springcloud/nacos/server/auth-service/src/main/java/com/dubbogo/pixiu/spring/cloud/NacosProviderApplication.java
