+++ samples/springcloud/nacos/server/auth-service/start.sh
