+++ samples/springcloud/nacos/server/user-service/src/main/resources/application.properties
