+++ samples/springcloud/zookeeper/server/server-zk/src/main/resources/application.yml
