+++ samples/springcloud/zookeeper/server/server-zk/start.sh
